Move benchmark to GeoDistanceAccuracyTest class
